backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
qdrant-client==1.16.1
psycopg[binary]==3.3.2
asyncpg==0.29.0
# openai-chatkit==0.0.1 # Commented out - package not available on PyPI
openai==1.3.8
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
email-validator==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Development Dependencies
black==23.11.0
ruff==0.1.7
mypy==1.7.0
isort==5.12.0
pre-commit==3.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
